backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.28.1
openai==1.58.1
PyPDF2==3.0.1
Pillow==12.0.0
pytest==8.3.4
pytest-asyncio==0.24.0
